requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
openai
alembic
jinja2
aiofiles
python-dotenv